Move flit source code to src/flit/
